Revert "refactor(gms): move gms restli resources (#1619)"
This reverts commit 6a309332f321fc6847a721629c276e62867bf52c.
