+++ gms/api/src/main/idl/com.linkedin.identity.corpGroups.restspec.json
@@ -3,7 +3,7 @@
   "namespace" : "com.linkedin.identity",
   "path" : "/corpGroups",
   "schema" : "com.linkedin.identity.CorpGroup",
-  "doc" : "generated from: com.linkedin.metadata.resources.identity.CorpGroups",
+  "doc" : "generated from: com.linkedin.identity.rest.resources.CorpGroups",
   "collection" : {
     "identifier" : {
       "name" : "corpGroup",
