+++ gms/api/src/main/snapshot/com.linkedin.dataset.datasets.snapshot.json
@@ -1156,7 +1156,7 @@
     "namespace" : "com.linkedin.dataset",
     "path" : "/datasets",
     "schema" : "com.linkedin.dataset.Dataset",
-    "doc" : "generated from: com.linkedin.metadata.resources.dataset.Datasets",
+    "doc" : "generated from: com.linkedin.dataset.rest.resources.Datasets",
     "collection" : {
       "identifier" : {
         "name" : "dataset",
@@ -1278,7 +1278,7 @@
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/deprecation",
           "schema" : "com.linkedin.dataset.DatasetDeprecation",
-          "doc" : "Rest.li entry point: /datasets/{datasetKey}/deprecation\n\ngenerated from: com.linkedin.metadata.resources.dataset.DeprecationResource",
+          "doc" : "Rest.li entry point: /datasets/{datasetKey}/deprecation\n\ngenerated from: com.linkedin.dataset.rest.resources.DeprecationResource",
           "collection" : {
             "identifier" : {
               "name" : "deprecationId",
@@ -1299,7 +1299,7 @@
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/downstreamLineage",
           "schema" : "com.linkedin.dataset.DownstreamLineage",
-          "doc" : "Rest.li entry point: /datasets/{datasetKey}/downstreamLineage\n\ngenerated from: com.linkedin.metadata.resources.dataset.DownstreamLineageResource",
+          "doc" : "Rest.li entry point: /datasets/{datasetKey}/downstreamLineage\n\ngenerated from: com.linkedin.dataset.rest.resources.DownstreamLineageResource",
           "simple" : {
             "supports" : [ "get" ],
             "methods" : [ {
@@ -1314,7 +1314,7 @@
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/institutionalMemory",
           "schema" : "com.linkedin.common.InstitutionalMemory",
-          "doc" : "Rest.li entry point: /datasets/{datasetKey}/institutionalMemory\n\ngenerated from: com.linkedin.metadata.resources.dataset.InstitutionalMemoryResource",
+          "doc" : "Rest.li entry point: /datasets/{datasetKey}/institutionalMemory\n\ngenerated from: com.linkedin.dataset.rest.resources.InstitutionalMemoryResource",
           "collection" : {
             "identifier" : {
               "name" : "institutionalMemoryId",
@@ -1335,7 +1335,7 @@
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/rawOwnership",
           "schema" : "com.linkedin.common.Ownership",
-          "doc" : "Rest.li entry point: /datasets/{datasetKey}/rawOwnership\n\ngenerated from: com.linkedin.metadata.resources.dataset.OwnershipResource",
+          "doc" : "Rest.li entry point: /datasets/{datasetKey}/rawOwnership\n\ngenerated from: com.linkedin.dataset.rest.resources.OwnershipResource",
           "collection" : {
             "identifier" : {
               "name" : "rawOwnershipId",
@@ -1356,7 +1356,7 @@
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/schema",
           "schema" : "com.linkedin.schema.SchemaMetadata",
-          "doc" : "Rest.li entry point: /datasets/{datasetKey}/schema\n\ngenerated from: com.linkedin.metadata.resources.dataset.SchemaResource",
+          "doc" : "Rest.li entry point: /datasets/{datasetKey}/schema\n\ngenerated from: com.linkedin.dataset.rest.resources.SchemaResource",
           "collection" : {
             "identifier" : {
               "name" : "schemaId",
@@ -1377,7 +1377,7 @@
           "namespace" : "com.linkedin.dataset",
           "path" : "/datasets/{dataset}/upstreamLineage",
           "schema" : "com.linkedin.dataset.UpstreamLineage",
-          "doc" : "Rest.li entry point: /datasets/{datasetKey}/upstreamLineage\n\ngenerated from: com.linkedin.metadata.resources.dataset.UpstreamLineageResource",
+          "doc" : "Rest.li entry point: /datasets/{datasetKey}/upstreamLineage\n\ngenerated from: com.linkedin.dataset.rest.resources.UpstreamLineageResource",
           "collection" : {
             "identifier" : {
               "name" : "upstreamLineageId",
