+++ gms/impl/build.gradle
@@ -24,8 +24,8 @@ dependencies {
 
 // Generate IDLs
 pegasus.main.idlOptions.addIdlItem([
-    'com.linkedin.metadata.resources.dataset',
-    'com.linkedin.metadata.resources.identity'
+    'com.linkedin.dataset.rest',
+    'com.linkedin.identity.rest'
 ])
 
 ext.apiProject = project(':gms:api')
