+++ gms/impl/src/main/java/com/linkedin/dataset/rest/resources/DownstreamLineageResource.java
@@ -1,4 +1,4 @@
-package com.linkedin.metadata.resources.dataset;
+package com.linkedin.dataset.rest.resources;
 
 import com.linkedin.common.urn.DatasetUrn;
 import com.linkedin.dataset.*;
