+++ gms/impl/src/main/java/com/linkedin/identity/rest/resources/CorpUsers.java
@@ -1,4 +1,4 @@
-package com.linkedin.metadata.resources.identity;
+package com.linkedin.identity.rest.resources;
 
 import com.linkedin.common.urn.CorpuserUrn;
 import com.linkedin.common.urn.Urn;
